+++ dev/openchpl_views.sql
@@ -356,6 +356,48 @@ WHERE (certification_body.deleted = false OR certification_body.deleted IS NULL)
 
 --ALTER TABLE openchpl.acb_developer_transparency_mappings OWNER TO openchpl;
 
+DROP VIEW IF EXISTS openchpl.developers_with_attestations;
+CREATE OR REPLACE VIEW openchpl.developers_with_attestations AS
+SELECT
+v.vendor_id as vendor_id,
+v.name as vendor_name,
+s.name as status_name,
+sum(case when certification_status.certification_status = 'Active' then 1 else 0 end) as countActiveListings,
+sum(case when certification_status.certification_status = 'Retired' then 1 else 0 end) as countRetiredListings,
+sum(case when certification_status.certification_status = 'Pending' then 1 else 0 end) as countPendingListings,
+sum(case when certification_status.certification_status = 'Withdrawn by Developer' then 1 else 0 end) as countWithdrawnByDeveloperListings,
+sum(case when certification_status.certification_status = 'Withdrawn by ONC-ACB' then 1 else 0 end) as countWithdrawnByOncAcbListings,
+sum(case when certification_status.certification_status = 'Suspended by ONC-ACB' then 1 else 0 end) as countSuspendedByOncAcbListings,
+sum(case when certification_status.certification_status = 'Suspended by ONC' then 1 else 0 end) as countSuspendedByOncListings,
+sum(case when certification_status.certification_status = 'Terminated by ONC' then 1 else 0 end) as countTerminatedByOncListings,
+sum(case when certification_status.certification_status = 'Withdrawn by Developer Under Surveillance/Review' then 1 else 0 end) as countWithdrawnByDeveloperUnderSurveillanceListings,
+
+--only include urls that are not empty strings and come from
+-- a listing with one of the active... or suspended... statuses
+string_agg(DISTINCT 
+	case when 
+		listings.transparency_attestation_url::text != '' 
+		and 
+			(certification_status.certification_status = 'Active' 
+			or
+			certification_status.certification_status = 'Suspended by ONC'
+			or 
+			certification_status.certification_status = 'Suspended by ONC-ACB')
+		then listings.transparency_attestation_url::text else null end, '☺') 
+	as "transparency_attestation_urls",
+--using coalesce here because the attestation can be null and concatting null with anything just gives null
+--so null/empty attestations are left out unless we replace null with empty string
+string_agg(DISTINCT acb.name::text||':'||COALESCE(attestations.transparency_attestation::text, ''), '☺') as "attestations"
+FROM openchpl.vendor v
+LEFT OUTER JOIN openchpl.vendor_status s ON v.vendor_status_id = s.vendor_status_id
+LEFT OUTER JOIN openchpl.certified_product_details listings ON listings.vendor_id = v.vendor_id AND listings.deleted != true
+LEFT OUTER JOIN openchpl.certification_status ON listings.certification_status_id = certification_status.certification_status_id
+LEFT OUTER JOIN openchpl.acb_vendor_map attestations ON attestations.vendor_id = v.vendor_id AND attestations.deleted != true
+LEFT OUTER JOIN openchpl.certification_body acb ON attestations.certification_body_id = acb.certification_body_id AND acb.deleted != true
+
+WHERE v.deleted != true
+GROUP BY v.vendor_id, v.name, s.name;
+
 CREATE OR REPLACE VIEW openchpl.ehr_certification_ids_and_products AS
 SELECT 
 	row_number() OVER () AS id,